requirements.txt
fastapi[all]
psycopg2
passlib[bcrypt]
sqlalchemy
pydantic
pyjwt